adding price validation + changing double to big decimal

+++ src/main/java/Main.java
@@ -1,6 +1,7 @@
 import domain.BasketItem;
 import domain.Item;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 
 public class Main {
@@ -9,11 +10,11 @@ public class Main {
 
         Basket basket = new Basket();
 
-        BasketItem banana = new BasketItem(Item.BANANA,0.20);
-        BasketItem orange = new BasketItem(Item.ORANGE,0.10);
-        BasketItem apple = new BasketItem(Item.APPLE,0.30);
-        BasketItem lemon = new BasketItem(Item.LEMON,0.24);
-        BasketItem peach = new BasketItem(Item.PEACH,0.31);
+        BasketItem banana = new BasketItem(Item.BANANA, BigDecimal.valueOf(0.20));
+        BasketItem orange = new BasketItem(Item.ORANGE, BigDecimal.valueOf(0.10));
+        BasketItem apple = new BasketItem(Item.APPLE, BigDecimal.valueOf(0.30));
+        BasketItem lemon = new BasketItem(Item.LEMON, BigDecimal.valueOf(0.24));
+        BasketItem peach = new BasketItem(Item.PEACH, BigDecimal.valueOf(0.31));
 
         basket.add(banana);
         basket.add(orange);
